refactor: move `withTempBlobData` to filesys
