+++ packages/blocks/src/attachment-block/index.ts
@@ -1,3 +1,2 @@
 export * from './attachment-block.js';
 export * from './attachment-model.js';
-export { withTempBlobData } from './utils.js';
